+++ R/containers.R
@@ -101,6 +101,7 @@ NULL
 #' @export
 iv_containers <- function(x) {
   proxy <- iv_proxy(x)
+  check_iv(proxy, arg = "x")
 
   start <- field_start(proxy)
   end <- field_end(proxy)