+++ R/relation.R
@@ -508,17 +508,30 @@ iv_pairwise_overlaps <- function(x,
   iv_detect_pairwise_impl(x, y, type, iv_prepare_overlaps)
 }
 
-iv_prepare_overlaps <- function(needles, haystack, type) {
-  type <- arg_match0(type, c("any", "equals", "contains", "within", "starts", "ends"))
+iv_prepare_overlaps <- function(needles,
+                                haystack,
+                                type,
+                                ...,
+                                error_call = caller_env()) {
+  check_dots_empty0(...)
 
-  needles_proxy <- iv_proxy(needles)
-  haystack_proxy <- iv_proxy(haystack)
+  type <- arg_match0(
+    arg = type,
+    values = c("any", "equals", "contains", "within", "starts", "ends"),
+    error_call = error_call
+  )
 
-  needles_start <- field_start(needles_proxy)
-  haystack_start <- field_start(haystack_proxy)
+  needles <- iv_proxy(needles)
+  check_iv(needles, call = error_call)
 
-  needles_end <- field_end(needles_proxy)
-  haystack_end <- field_end(haystack_proxy)
+  haystack <- iv_proxy(haystack)
+  check_iv(haystack, call = error_call)
+
+  needles_start <- field_start(needles)
+  haystack_start <- field_start(haystack)
+
+  needles_end <- field_end(needles)
+  haystack_end <- field_end(haystack)
 
   if (type == "any") {
     needles <- data_frame(a = needles_start, b = needles_end)
@@ -625,10 +638,7 @@ iv_locate_positional <- function(needles,
   needles <- args[[1L]]
   haystack <- args[[2L]]
 
-  needles <- iv_proxy(needles)
-  haystack <- iv_proxy(haystack)
-
-  args <- iv_prepare_positional(needles, haystack, type)
+  args <- iv_prepare_positional(needles, haystack, type, error_call = error_call)
   needles <- args$needles
   haystack <- args$haystack
   condition <- args$condition
@@ -827,15 +837,24 @@ iv_detect_pairwise_positional <- function(x,
 }
 
 
-iv_prepare_positional <- function(needles, haystack, type) {
-  needles_proxy <- iv_proxy(needles)
-  haystack_proxy <- iv_proxy(haystack)
+iv_prepare_positional <- function(needles,
+                                  haystack,
+                                  type,
+                                  ...,
+                                  error_call = caller_env()) {
+  check_dots_empty0(...)
 
-  needles_start <- field_start(needles_proxy)
-  haystack_start <- field_start(haystack_proxy)
+  needles <- iv_proxy(needles)
+  check_iv(needles, call = error_call)
 
-  needles_end <- field_end(needles_proxy)
-  haystack_end <- field_end(haystack_proxy)
+  haystack <- iv_proxy(haystack)
+  check_iv(haystack, call = error_call)
+
+  needles_start <- field_start(needles)
+  haystack_start <- field_start(haystack)
+
+  needles_end <- field_end(needles)
+  haystack_end <- field_end(haystack)
 
   if (type == "precedes") {
     needles <- data_frame(a = needles_end)
@@ -1326,10 +1345,16 @@ iv_pairwise_relates <- function(x, y, ..., type) {
   iv_detect_pairwise_impl(x, y, type, iv_prepare_relation)
 }
 
-iv_prepare_relation <- function(needles, haystack, type) {
+iv_prepare_relation <- function(needles,
+                                haystack,
+                                type,
+                                ...,
+                                error_call = caller_env()) {
+  check_dots_empty0(...)
+
   type <- arg_match0(
-    type,
-    c(
+    arg = type,
+    values = c(
       "precedes",
       "preceded-by",
       "meets",
@@ -1343,17 +1368,21 @@ iv_prepare_relation <- function(needles, haystack, type) {
       "during",
       "contains",
       "equals"
-    )
+    ),
+    error_call = error_call
   )
 
-  needles_proxy <- iv_proxy(needles)
-  haystack_proxy <- iv_proxy(haystack)
+  needles <- iv_proxy(needles)
+  check_iv(needles, call = error_call)
 
-  needles_start <- field_start(needles_proxy)
-  haystack_start <- field_start(haystack_proxy)
+  haystack <- iv_proxy(haystack)
+  check_iv(haystack, call = error_call)
 
-  needles_end <- field_end(needles_proxy)
-  haystack_end <- field_end(haystack_proxy)
+  needles_start <- field_start(needles)
+  haystack_start <- field_start(haystack)
+
+  needles_end <- field_end(needles)
+  haystack_end <- field_end(haystack)
 
   if (type == "precedes") {
     needles <- data_frame(a = needles_end)
@@ -1476,7 +1505,7 @@ iv_detect_impl <- function(needles,
   needles <- args[[1L]]
   haystack <- args[[2L]]
 
-  args <- iv_prepare_impl(needles, haystack, type)
+  args <- iv_prepare_impl(needles, haystack, type, error_call = error_call)
   needles <- args$needles
   haystack <- args$haystack
   condition <- args$condition
@@ -1537,7 +1566,7 @@ iv_detect_pairwise_impl <- function(x,
   x <- args[[1L]]
   y <- args[[2L]]
 
-  args <- iv_prepare_impl(x, y, type)
+  args <- iv_prepare_impl(x, y, type, error_call = error_call)
   # https://github.com/r-lib/rlang/issues/1346
   args <- map(args, unname)
   args <- transpose(args)